requirements.txt
selenium>=4.0.0
pytest>=7.0.0
python-dotenv
requests
browserstack-sdk 